Move ContentNodeGranular into its own module
The three ValidationErrors resolved against kolibri.core.content.serializers,
which has no such attribute, so the multi-flag and bad-drive/peer paths raised
AttributeError and 500ed. In the new module the name is rest_framework's and
they raise the intended 400.
