+++ kolibri/core/content/api_urls.py
@@ -4,7 +4,6 @@ from django.urls import re_path
 from rest_framework import routers
 
 from .api import ChannelThumbnailView
-from .api import ContentNodeGranularViewset
 from .api import FileViewset
 from .api import RemoteChannelViewSet
 from .api import ShareFileView
@@ -12,6 +11,7 @@ from .viewsets.channel_metadata import ChannelMetadataViewSet
 from .viewsets.content_request import ContentRequestViewset
 from .viewsets.contentnode.base import ContentNodeViewset
 from .viewsets.contentnode.bookmarks import ContentNodeBookmarksViewset
+from .viewsets.contentnode.granular import ContentNodeGranularViewset
 from .viewsets.contentnode.progress import ContentNodeProgressViewset
 from .viewsets.contentnode.tree import ContentNodeTreeViewset
 from .viewsets.contentnode.user import UserContentNodeViewset
